requirements.txt
Flask==2.0.2
Pillow==9.0.1